+++ src/main/java/com/botmaker/plugin/toolkit/Editors.java
@@ -1,6 +1,6 @@
 package com.botmaker.plugin.toolkit;
 
-import com.botmaker.plugin.api.ValueContext;
+import com.botmaker.plugin.api.slot.ValueContext;
 import javafx.scene.Node;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.ComboBox;