+++ src/test/java/com/botmaker/plugin/toolkit/CallSitesTest.java
@@ -1,6 +1,6 @@
 package com.botmaker.plugin.toolkit;
 
-import com.botmaker.plugin.api.ValueContext;
+import com.botmaker.plugin.api.slot.ValueContext;
 import com.botmaker.plugin.toolkit.testing.TestContexts;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;